requirements/base.txt
fastapi==0.103.1
loguru==0.7.1
SQLAlchemy==2.0.20
alembic==1.12.0
asyncpg==0.28.0
greenlet==2.0.2
python-multipart==0.0.6
aiofiles==23.2.1
python-dotenv==1.0.0
uvicorn==0.23.2